更换后台amazeui 为h-ui，添加未读消息数量显示

+++ Nutshell.Blog.IReposotory/IModuleRepository.cs
@@ -9,7 +9,7 @@ namespace Nutshell.Blog.IReposotory
 {
     public interface IModuleRepository : IBaseRepository<Module>
     {
-        List<Module> GetMenuByPersonId(int personId);
+        List<Module> GetMenuByPersonId(long personId);
 
         //IQueryable<Module> GetModuleList();
     }
